refactored according to newest standards
